fix(api): fix tests with wrong model
