+++ api/tests/devcomp/unit/domain/models/element/QCU-declarative_test.js
@@ -1,5 +1,8 @@
+import { ModuleInstantiationError } from '../../../../../../src/devcomp/domain/errors.js';
 import { QCUDeclarative } from '../../../../../../src/devcomp/domain/models/element/QCU-declarative.js';
+import { DomainError } from '../../../../../../src/shared/domain/errors.js';
 import { expect } from '../../../../../test-helper.js';
+import { catchErrSync } from '../../../../../tooling/test-utils/error.js';
 
 describe('Unit | Devcomp | Domain | Models | Element | QCU-declarative', function () {
   describe('#constructor', function () {
@@ -22,6 +25,50 @@ describe('Unit | Devcomp | Domain | Models | Element | QCU-declarative', functio
       expect(qcu.proposals).deep.equal([proposal1, proposal2]);
       expect(qcu.isAnswerable).to.be.true;
       expect(qcu.hasShortProposals).to.be.false;
+    });
+  });
+
+  describe('A QCU Declarative without id', function () {
+    it('should throw an error', function () {
+      // when
+      const error = catchErrSync(() => new QCUDeclarative({}))();
+
+      // then
+      expect(error).to.be.instanceOf(DomainError);
+      expect(error.message).to.equal('The id is required for an element');
+    });
+  });
+
+  describe('A QCUDeclarative without instruction', function () {
+    it('should throw an error', function () {
+      // when
+      const error = catchErrSync(() => new QCUDeclarative({ id: '123' }))();
+
+      // then
+      expect(error).to.be.instanceOf(DomainError);
+      expect(error.message).to.equal('The instruction is required for a QCU declarative');
+    });
+  });
+
+  describe('A QCUDeclarative with an empty list of proposals', function () {
+    it('should throw an error', function () {
+      // when
+      const error = catchErrSync(() => new QCUDeclarative({ id: '123', instruction: 'toto', proposals: [] }))();
+
+      // then
+      expect(error).to.be.instanceOf(ModuleInstantiationError);
+      expect(error.message).to.equal('The proposals are required for a QCU declarative');
+    });
+  });
+
+  describe('A QCUDeclarative does not have a list of proposals', function () {
+    it('should throw an error', function () {
+      // when
+      const error = catchErrSync(() => new QCUDeclarative({ id: '123', instruction: 'toto', proposals: 'toto' }))();
+
+      // then
+      expect(error).to.be.instanceOf(ModuleInstantiationError);
+      expect(error.message).to.equal('The QCU declarative proposals should be a list');
     });
   });
 });
